python/iot3: use test.mosquitto.org as a broker for tests
In the future, we'll be adding support for TLS and WebSockets, which
require some setup that a local default installation of an MQTT broker
will most probably not have (e.g. a server certificate...)

The test broker at test.moquitto.org exposes all kinds of setups, so it
is easy to test all of that: TLS, WebSockets, etc...

Switch to using that broker for our tests.

Randomise the client ID and topic to decrease the risk of a collision
with other clients.

Signed-off-by: Yann E. MORIN <yann.morin@orange.com>